# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HealthManager API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/appointments/{id}": {
      "patch": {
        "operationId": "appointments_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/appointment-types": {
      "get": {
        "operationId": "appointmentTypes_list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedAppointmentTypeResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "appointmentTypes_create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentTypeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentTypeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/appointment-types/{id}": {
      "put": {
        "operationId": "appointmentTypes_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentTypeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentTypeResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "appointmentTypes_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/financial/receivables": {
      "get": {
        "operationId": "receivables_list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Pending",
                "Partial",
                "Paid"
              ]
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedReceivableResponse"
                }
              }
            }
          }
        }
      }
    },
    "/financial/receivables/manual": {
      "post": {
        "operationId": "receivables_manual",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManualReceivableRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceivableResponse"
                }
              }
            }
          }
        }
      }
    },
    "/financial/payments": {
      "get": {
        "operationId": "payments_list",
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "ReceivableId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "DateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "DateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "DestinationBank",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedPaymentResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "payments_create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/appointments/{appointmentId}/clinical-record": {
      "get": {
        "operationId": "clinical_record_get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicalRecordResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "operationId": "clinical_record_create",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClinicalRecordRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicalRecordResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "clinical_record_update",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClinicalRecordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicalRecordResponse"
                }
              }
            }
          }
        }
      }
    },
    "/appointments/{appointmentId}/clinical-record/finalize": {
      "post": {
        "operationId": "clinical_record_finalize",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicalRecordResponse"
                }
              }
            }
          }
        }
      }
    },
    "/appointments/{appointmentId}/clinical-record/addendum": {
      "post": {
        "operationId": "clinical_record_add_addendum",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddendumRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicalRecordAddendumResponse"
                }
              }
            }
          }
        }
      }
    },
    "/appointments/{appointmentId}/clinical-record/addendums": {
      "get": {
        "operationId": "clinical_record_list_addendums",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClinicalRecordAddendumResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/financial/professional-settlements": {
      "get": {
        "operationId": "professional_settlements_list",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfessionalSettlementResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "professional_settlements_create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfessionalSettlementRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementResponse"
                }
              }
            }
          }
        }
      }
    },
    "/financial/owner-settlements": {
      "post": {
        "operationId": "owner_settlements_create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerSettlementRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OwnerSettlementRequest": {
        "type": "object",
        "properties": {
          "throughDate": {
            "type": "string",
            "format": "date"
          },
          "settledAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PagedAppointmentTypeResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppointmentTypeResponse"
            }
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "AppointmentTypeRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          }
        }
      },
      "AppointmentTypeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "PagedReceivableResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReceivableResponse"
            }
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
...
      "PagedPaymentResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentResponse"
            }
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "UpdateAppointmentRequest": {
        "type": "object",
        "properties": {
          "doctorId": {
            "type": "string",
            "format": "uuid"
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "durationMinutes": {
            "type": "integer"
          },
          "notes": {
            "type": "string"
          },
          "appointmentTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          }
        }
      },
...
      "AppointmentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "doctorId": {
            "type": "string",
            "format": "uuid"
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "confirmationStatus": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "appointmentTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "notes": {
            "type": "string"
          },
          "patientName": {
            "type": "string"
          },
          "patientPhone": {
            "type": "string"
          },
          "doctorName": {
            "type": "string"
          },
          "doctorSpecialty": {
            "type": "string"
          }
        }
      },
      "ReceivableResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "appointmentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "originalAmount": {
            "type": "number",
            "format": "decimal"
          },
          "receivedAmount": {
            "type": "number",
            "format": "decimal"
          },
          "outstandingAmount": {
            "type": "number",
            "format": "decimal"
          },
          "status": {
            "type": "string"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "patientName": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "appointmentStartAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "appointmentType": {
            "type": "string",
            "nullable": true
          },
          "appointmentStatus": {
            "type": "string",
            "nullable": true
          },
          "doctorId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "doctorName": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreatePaymentRequest": {
        "type": "object",
        "required": [
          "receivableId",
          "amount",
          "paymentMethod"
        ],
        "properties": {
          "receivableId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "paymentMethod": {
            "type": "string"
          },
          "paidAt": {
            "type": "string",
            "format": "date-time"
          },
          "destinationBank": {
            "type": "string",
            "maxLength": 100,
            "nullable": true
          },
          "fundsRecipient": {
            "type": "string",
            "enum": [
              "Clinic",
              "Owner"
            ],
            "default": "Clinic"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "CreateManualReceivableRequest": {
        "type": "object",
        "required": [
          "amount",
          "paymentMethod"
        ],
        "properties": {
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "description": {
            "type": "string"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "paymentMethod": {
            "type": "string"
          },
          "paidAt": {
            "type": "string",
            "format": "date-time"
          },
          "destinationBank": {
            "type": "string",
            "maxLength": 100,
            "nullable": true
          },
          "fundsRecipient": {
            "type": "string",
            "enum": [
              "Clinic",
              "Owner"
            ],
            "default": "Clinic"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "PaymentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "receivableId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "paymentMethod": {
            "type": "string"
          },
          "paidAt": {
            "type": "string",
            "format": "date-time"
          },
          "patientName": {
            "type": "string"
          },
          "destinationBank": {
            "type": "string",
            "nullable": true
          },
          "fundsRecipient": {
            "type": "string"
          },
          "clinicRevenueAmount": {
            "type": "number",
            "format": "decimal"
          },
          "professionalPayableAmount": {
            "type": "number",
            "format": "decimal"
          }
        }
      },
...
      "CreateClinicalRecordRequest": {
        "type": "object",
        "properties": {
          "chiefComplaint": {
            "type": "string",
            "nullable": true
          },
          "history": {
            "type": "string",
            "nullable": true
          },
          "physicalExam": {
            "type": "string",
            "nullable": true
          },
          "assessment": {
            "type": "string",
            "nullable": true
          },
          "plan": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateClinicalRecordRequest": {
        "type": "object",
        "properties": {
          "chiefComplaint": {
            "type": "string",
            "nullable": true
          },
          "history": {
            "type": "string",
            "nullable": true
          },
          "physicalExam": {
            "type": "string",
            "nullable": true
          },
          "assessment": {
            "type": "string",
            "nullable": true
          },
          "plan": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ClinicalRecordResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "appointmentId": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "doctorId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "Draft",
              "Finalized"
            ]
          },
          "chiefComplaint": {
            "type": "string",
            "nullable": true
          },
          "history": {
            "type": "string",
            "nullable": true
          },
          "physicalExam": {
            "type": "string",
            "nullable": true
          },
          "assessment": {
            "type": "string",
            "nullable": true
          },
          "plan": {
            "type": "string",
            "nullable": true
          },
          "finalizedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "patientName": {
            "type": "string",
            "nullable": true
          },
          "doctorName": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateAddendumRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
      "ClinicalRecordAddendumResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": "string"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          },
          "authorName": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}